feat: include the most recent message in each chat
