+++ src/components/PrivateChats/_privateChats.scss
@@ -13,7 +13,7 @@
   .chats-list {
     display: flex;
     flex-direction: column;
-    gap: 10px;
+    gap: 15px;
   }
   .add-friend {
     display: flex;
